Move mapping functions into the service
